+++ src/main/java/com/atlauncher/utils/Utils.java
@@ -2158,61 +2158,6 @@ public class Utils {
         return getMD5(returnStr);
     }
 
-    /**
-     * Credit to
-     * https://github.com/Slowpoke101/FTBLaunch/blob/master/src/main/java/net/ftb/workers/AuthlibDLWorker.java
-     */
-    public static boolean addToClasspath(File file) {
-        LogManager.info("Loading external library " + file.getName() + " to classpath");
-        try {
-            if (file.exists()) {
-                addURL(file.toURI().toURL());
-            } else {
-                LogManager.error("Error loading AuthLib");
-            }
-        } catch (Throwable t) {
-            if (t.getMessage() != null) {
-                LogManager.error(t.getMessage());
-            }
-            return false;
-        }
-
-        return true;
-    }
-
-    public static boolean checkAuthLibLoaded() {
-        try {
-            App.settings.getClass().forName("com.mojang.authlib.exceptions.AuthenticationException");
-            App.settings.getClass().forName("com.mojang.authlib.Agent");
-            App.settings.getClass().forName("com.mojang.authlib.yggdrasil.YggdrasilAuthenticationService");
-            App.settings.getClass().forName("com.mojang.authlib.yggdrasil.YggdrasilUserAuthentication");
-        } catch (ClassNotFoundException e) {
-            LogManager.logStackTrace(e);
-            return false;
-        }
-
-        return true;
-    }
-
-    /**
-     * Credit to
-     * https://github.com/Slowpoke101/FTBLaunch/blob/master/src/main/java/net/ftb/workers/AuthlibDLWorker.java
-     */
-    public static void addURL(URL u) throws IOException {
-        URLClassLoader sysloader = (URLClassLoader) App.settings.getClass().getClassLoader();
-        Class sysclass = URLClassLoader.class;
-        try {
-            Method method = sysclass.getDeclaredMethod("addURL", URL.class);
-            method.setAccessible(true);
-            method.invoke(sysloader, u);
-        } catch (Throwable t) {
-            if (t.getMessage() != null) {
-                LogManager.error(t.getMessage());
-            }
-            throw new IOException("Error, could not add URL to system classloader");
-        }
-    }
-
     public static String convertMavenIdentifierToPath(String identifier) {
         String[] parts = identifier.split(":", 3);
         String name = parts[1];