+++ CHANGELOG.md
@@ -4,3 +4,4 @@
 
 -   Fix issue with skins not downloading causes application hang
 -   Fix issue with some Forge loader libraries not installing for servers
+-   Update authlib version and remove classloading hack