+++ CHANGELOG.md
@@ -6,3 +6,4 @@
 -   Fix issue with some Forge loader libraries not installing for servers
 -   Update authlib version and remove classloading hack
 -   Remove deprecated system information api call
+-   Fix network checker not working